Tensorboard

+++ dataset/dataset.py
@@ -438,10 +438,10 @@ def loadDatasetRaw(numPairs):
 
 	newDatasets[0] = newDatasets[0][:numPairs]
 	newDatasets[1] = newDatasets[1][:numPairs]
-	newDatasets[2] = newDatasets[2][:numPairs / 4]
-	newDatasets[3] = newDatasets[3][:numPairs / 4]
-	newDatasets[4] = newDatasets[4][:numPairs / 4]
-	newDatasets[5] = newDatasets[5][:numPairs / 4]
+	newDatasets[2] = newDatasets[2][:int(numPairs / 4)]
+	newDatasets[3] = newDatasets[3][:int(numPairs / 4)]
+	newDatasets[4] = newDatasets[4][:int(numPairs / 4)]
+	newDatasets[5] = newDatasets[5][:int(numPairs / 4)]
 	
 	return newDatasets
